tech(api): moving incrementInContext from monitoring-tools to execution-context-manager
